[types] move epoch info into its own module
No point in having this in crypto proxies except that it was shortcutted
to take advantage of the types defined therein.
