+++ consensus/src/chained_bft/event_processor.rs
@@ -43,7 +43,7 @@ use libra_crypto::hash::TransactionAccumulatorHasher;
 use libra_logger::prelude::*;
 use libra_security_logger::{security_log, SecurityEvent};
 use libra_types::{
-    crypto_proxies::EpochInfo, ledger_info::LedgerInfoWithSignatures,
+    epoch_info::EpochInfo, ledger_info::LedgerInfoWithSignatures,
     transaction::TransactionStatus, validator_change::ValidatorChangeProof,
     validator_verifier::ValidatorVerifier,
 };
