+++ state-synchronizer/src/lib.rs
@@ -12,7 +12,7 @@ extern crate prometheus;
 
 use executor_types::ExecutedTrees;
 use libra_types::{
-    account_address::AccountAddress, crypto_proxies::EpochInfo,
+    account_address::AccountAddress, epoch_info::EpochInfo,
     ledger_info::LedgerInfoWithSignatures, validator_verifier::ValidatorVerifier,
 };
 use std::sync::Arc;
