+++ types/src/crypto_proxies.rs
@@ -21,7 +21,7 @@
 
 use crate::{
     validator_info::ValidatorInfo as RawValidatorInfo,
-    validator_set::ValidatorSet as RawValidatorSet, validator_verifier::ValidatorVerifier,
+    validator_set::ValidatorSet as RawValidatorSet,
 };
 
 // This sets the types containing cryptographic materials used in the
@@ -36,35 +36,6 @@ use crate::{
 // below is banned.
 
 use libra_crypto::ed25519::Ed25519PublicKey;
-use std::{collections::BTreeMap, fmt};
 
 pub type ValidatorInfo = RawValidatorInfo<Ed25519PublicKey>;
 pub type ValidatorSet = RawValidatorSet<Ed25519PublicKey>;
-use std::sync::Arc;
-
-#[derive(Clone, Debug)]
-/// EpochInfo represents a trusted validator set to validate messages from the specific epoch,
-/// it could be updated with ValidatorChangeProof.
-pub struct EpochInfo {
-    pub epoch: u64,
-    pub verifier: Arc<ValidatorVerifier>,
-}
-
-impl EpochInfo {
-    pub fn empty() -> Self {
-        Self {
-            epoch: 0,
-            verifier: Arc::new(ValidatorVerifier::new(BTreeMap::new())),
-        }
-    }
-}
-
-impl fmt::Display for EpochInfo {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(
-            f,
-            "EpochInfo [epoch: {}, validator: {}]",
-            self.epoch, self.verifier
-        )
-    }
-}
